+++ packages/components/condition/src/ConditionStep/ConditionStep.tsx
@@ -40,7 +40,6 @@ export const ConditionStep: React.FC<T.ConditionStepProps> = ({
   onDeactivate,
   showSuffix,
   hoverDisabled,
-  errorId,
 }) => {
   const { formatMessage } = useIntl();
   const text = React.useMemo(
@@ -161,7 +160,6 @@ export const ConditionStep: React.FC<T.ConditionStepProps> = ({
           texts={text}
           stepType={step.context?.type}
           onDeactivate={onDeactivate}
-          error={errorId === condition.id && !condition.factor.value && true}
         />
       );
     },
@@ -188,7 +186,6 @@ export const ConditionStep: React.FC<T.ConditionStepProps> = ({
       setCurrentField,
       setCurrentCondition,
       setCurrentStep,
-      errorId,
     ]
   );