+++ packages/components/date-range-picker/src/DateRangePicker.types.ts
@@ -5,6 +5,29 @@ import { DateFilter, DateRange, RelativeUnits } from './date.types';
 import { FilterDefinition } from './RangeFilter/RangeFilter.types';
 import { SavedFilter } from './RangeFilter/Shared/FilterDropdown/FilterDropdown.types';
 
+export type CustomColorArrow =
+  | 'blue'
+  | 'grey'
+  | 'red'
+  | 'green'
+  | 'yellow'
+  | 'pink'
+  | 'mars'
+  | 'orange'
+  | 'fern'
+  | 'cyan'
+  | 'purple'
+  | 'violet';
+export type AdditionalMapper = {
+  topLeft: CustomColorArrow;
+  topRight: CustomColorArrow;
+  bottomLeft: CustomColorArrow;
+  bottomRight: CustomColorArrow;
+  leftTop: CustomColorArrow;
+  leftBottom: CustomColorArrow;
+  rightTop: CustomColorArrow;
+  rightBottom: CustomColorArrow;
+};
 export interface Props extends WrappedComponentProps {
   disableDefaultTexts?: boolean;
   disabledDate?: (date?: Date) => boolean;
@@ -30,6 +53,7 @@ export interface Props extends WrappedComponentProps {
   texts: Texts;
   validate?: (value: DateRange) => { valid: boolean; message?: string };
   value: DateRange;
+  arrowColor?: AdditionalMapper;
 }
 
 export type RelativeMode = 'PAST' | 'FUTURE' | 'SINCE';