+++ packages/components/date-range-picker/src/RawDateRangePicker.tsx
@@ -1,9 +1,10 @@
 import * as React from 'react';
 import { omitBy, isUndefined } from 'lodash';
 import { injectIntl } from 'react-intl';
+import fnsIsValid from 'date-fns/isValid';
 import { Container, Separator, Addon } from './DateRangePicker.styles';
 import RangePicker from './RangePicker/RangePicker';
-import { RELATIVE, ABSOLUTE, MODES, RELATIVE_PRESETS } from './constants';
+import { RELATIVE, ABSOLUTE, MODES, RELATIVE_PRESETS, ABSOLUTE_PRESETS } from './constants';
 import relativeToAbsolute from './dateUtils/relativeToAbsolute';
 import { Props, State, AddonType } from './DateRangePicker.types';
 import { DateFilter, DateRange } from './date.types';
@@ -15,9 +16,9 @@ import RangeFilter from './RangeFilter/RangeFilter';
 import RangeFilterStatus from './RangeFilter/Shared/RangeFilterStatus/RangeFilterStatus';
 import { FilterDefinition, FilterValue } from './RangeFilter/RangeFilter.types';
 
-class RawDateRangePicker extends React.PureComponent<Props, State> {
+export class RawDateRangePicker extends React.PureComponent<Props, State> {
   static defaultProps = {
-    ranges: RELATIVE_PRESETS,
+    ranges: [...RELATIVE_PRESETS, ...ABSOLUTE_PRESETS],
     relativePast: true,
     showRelativePicker: true,
     validate: (): { valid: boolean } => ({ valid: true }),
@@ -38,10 +39,6 @@ class RawDateRangePicker extends React.PureComponent<Props, State> {
     if (prevProps.value !== value && !value) {
       this.handleRangeChange(value);
     }
-    const { mode } = this.state;
-    if (!value?.to && !value?.from && mode !== MODES.DATE) {
-      this.handleSwitchMode();
-    }
   }
 
   handleFilterCancel = (): void => {
@@ -57,10 +54,17 @@ class RawDateRangePicker extends React.PureComponent<Props, State> {
     if (!range) {
       return;
     }
+    if (range.from && !fnsIsValid(range?.from)) {
+      return;
+    }
 
     const { onValueChange } = this.props;
-    const { value } = this.state;
+    const { value, mode } = this.state;
     const newValue = normalizeRange({ ...range, filter: value.filter });
+    if (newValue.type === 'RELATIVE' && mode === MODES.TIME) {
+      // clicked on RangeButtons and was selecting time
+      this.setState({ mode: MODES.DATE });
+    }
     this.setState({ value: { ...newValue, key: range?.key, translationKey: range?.translationKey } });
     onValueChange && onValueChange(newValue);
   };
@@ -180,17 +184,8 @@ class RawDateRangePicker extends React.PureComponent<Props, State> {
   };
 
   render(): JSX.Element {
-    const {
-      showTime,
-      format,
-      disabledDate,
-      validate,
-      forceAdjacentMonths,
-      texts,
-      savedFilters,
-      onFilterSave,
-      intl,
-    } = this.props;
+    const { showTime, format, disabledDate, validate, forceAdjacentMonths, texts, savedFilters, onFilterSave, intl } =
+      this.props;
     const { value, mode } = this.state;
     const { from, to, key } = value;
     const addons = this.getAddons();