+++ packages/components/dropdown/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.18](https://github.com/synerise/synerise-design/compare/@synerise/ds-dropdown@0.17.17...@synerise/ds-dropdown@0.17.18) (2022-06-15)
+
+**Note:** Version bump only for package @synerise/ds-dropdown
+
+
+
+
+
 ## [0.17.17](https://github.com/synerise/synerise-design/compare/@synerise/ds-dropdown@0.17.16...@synerise/ds-dropdown@0.17.17) (2022-04-29)