+++ packages/components/grid/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.11](https://github.com/Synerise/synerise-design/compare/@synerise/ds-grid@0.3.10...@synerise/ds-grid@0.3.11) (2022-06-15)
+
+**Note:** Version bump only for package @synerise/ds-grid
+
+
+
+
+
 ## [0.3.10](https://github.com/Synerise/synerise-design/compare/@synerise/ds-grid@0.3.9...@synerise/ds-grid@0.3.10) (2022-04-29)
 
 **Note:** Version bump only for package @synerise/ds-grid